+++ src/app/navigationcontroller.h
@@ -19,8 +19,8 @@ class NavigationController
     Q_GADGET
 public:
     Q_INVOKABLE void showOnMap(const QVariant &place);
-    Q_INVOKABLE void showOnMap(float latitude, float longitude, int zoom);
-    Q_INVOKABLE void showOnWheelmap(float latitude, float longitude);
+    Q_INVOKABLE void showOnMap(double latitude, double longitude, int zoom);
+    Q_INVOKABLE void showOnWheelmap(double latitude, double longitude);
 
     Q_INVOKABLE bool canNavigateTo(const QVariant &place);
     Q_INVOKABLE void navigateTo(const QVariant &place);
